+++ internal/app/grpc/app.go
@@ -20,25 +20,6 @@ type App struct {
 	port       int
 }
 
-func maskSensitiveFields(fields []any) []any {
-	for i := 0; i < len(fields)-1; i += 2 {
-		key, ok := fields[i].(string)
-		if !ok {
-			continue
-		}
-		if key == "password" || key == "pwd" || key == "pass" {
-			if pwd, ok := fields[i+1].(string); ok {
-				if len(pwd) > 4 {
-					fields[i+1] = pwd[:1] + "****" + pwd[len(pwd)-1:]
-				} else {
-					fields[i+1] = "****"
-				}
-			}
-		}
-	}
-	return fields
-}
-
 func InterceptorLogger(l *slog.Logger) logging.Logger {
 	return logging.LoggerFunc(func(ctx context.Context, lvl logging.Level, msg string, fields ...any) {
 		maskedFields := maskSensitiveFields(fields)
